requirements.txt
pandas
openpyxl
watchdog
jinja2
python-dotenv